requirements.txt
pandas 
python-dotenv 
requests